+++ Chess/src/model/pieces/Knight.java
@@ -19,8 +19,8 @@ public class Knight extends Piece {
     @Override
     public void calculatePreliminaryMoves() {
         legalMoves.clear();
-        int r = (int) position.getX();
-        int c = (int) position.getY();
+        int r =  position.getX();
+        int c =  position.getY();
         ArrayList<Point> possibleMoves = new ArrayList<>();
 
         // down right
